backend/api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlmodel==0.0.14
python-dotenv==1.0.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fastapi-users[sqlalchemy]==12.1.2
aiohttp==3.9.1
pydantic==2.6.1
pydantic-settings==2.1.0 